Upgrade hugo and lots of npm packages
1. Upgrade hugo version to v0.56.3
2. Replace .URL with new link variables(e.g. .RelPermalink) in hugo
3. Remove useless theme.toml
4. Make content/about-cn/recruit as a section in hugo
5. Filter TODO.md files in hugo config.toml
6. Upgrade gulp version to v4.0.2
7. Use gulpfile.js, remove @babel/register
8. Split babel config and eslint config to a single file
9. Rename webpack config file
10. Add .prettierrc to better support code formatting
