+++ .eslintignore
@@ -1,8 +1,7 @@
 # /node_modules/* and /bower_components/* ignored by default
 
-# Ignore built files except dist/index.js
+# Ignore built files
 dist/*
-!dist/index.js
 
 # Ignore vendor files
 static/js/*
