+++ .eslintrc.json
@@ -1,0 +1,19 @@
+{
+  "extends": "prettier",
+  "plugins": ["prettier", "import"],
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "sourceType": "module"
+  },
+  "rules": {
+    "prettier/prettier": [
+      "error",
+      {
+        "singleQuote": true,
+        "semi": false,
+        "trailingComma": "es5"
+      }
+    ],
+    "no-script-url": 0
+  }
+}
