+++ config.toml
@@ -8,21 +8,23 @@ preserveTaxonomyNames = true
 pluralizeListTitles = true
 enableEmoji = true
 
-#Set to true to enable syntax highlighting in code fences with a language tag in markdown
+# Set to true to enable syntax highlighting in code fences with a language tag in markdown
 pygmentsCodefences = true
 # color-codes for highlighting derived from this style
 pygmentsStyle = "monokai"
 # true: use pygments-css or false: color-codes directly
 pygmentsUseClasses = false
-#Set to true to try to do syntax highlighting on code fenced blocks in markdown without a language tag
+# Set to true to try to do syntax highlighting on code fenced blocks in markdown without a language tag
 pygmentsCodeFencesGuessSyntax = true
-#Set to true to use Pygments instead of the much faster Chroma
+# Set to true to use Pygments instead of the much faster Chroma
 # pygmentsUseClassic = false
 
 # RSS, categories and tags disabled for an easy start
 # See configuration options for more details:
 # https://gohugo.io/getting-started/configuration/#toml-configuration
 disableKinds = ["taxonomy", "taxonomyTerm"]
+
+ignoreFiles = ["TODO\\.md$"]
 
 [params.versions.docs]
 total_versions = ['dev', 'v3.0', 'v2.1']
