+++ layouts/docs/list.html
@@ -16,7 +16,7 @@
 {{ $total_versions := .Site.Params.versions.docs.total_versions }}
 <div class="doc" id="list_page">
     <div class="container">
-        {{ $current_version := index (findRE "\\/(v\\d+\\.\\d+|dev)" .URL) 0 }}
+        {{ $current_version := index (findRE "\\/(v\\d+\\.\\d+|dev)" .RelPermalink) 0 }}
         
         {{- if eq $current_version "/dev" -}}
             {{ partial "archives/sticky-tree.html" (dict "tree" .Site.Data.docs_dev_toc "title" "Documentation") }}
@@ -35,7 +35,7 @@
                     {{ .Scratch.Add "data_i18_link" $current_version }}
                 {{ end }}
 
-                {{ $trimed_URL := strings.TrimPrefix (.Scratch.Get "prefix") .URL }}
+                {{ $trimed_URL := strings.TrimPrefix (.Scratch.Get "prefix") .RelPermalink }}
 
                 {{ $.Scratch.Set "doesnot_exist_ver" (slice "") }}
