+++ layouts/partials/archives/article-reading-card.html
@@ -4,7 +4,7 @@
 
 <div class="list article-list">
     {{- range .posts -}}
-    <a href="{{ .URL }}">
+    <a href="{{ .RelPermalink }}">
       <div class="blog__article" data-category="{{ range .Params.categories }}{{ replace . " " "-" }} {{ end }}">
           <!-- {{ printf "%#v" $listType }} -->
           {{- if ne $listType "blog" -}}
