+++ layouts/partials/archives/post-share-cn.html
@@ -3,7 +3,7 @@
         <a class="ssba_mail_share" href="mailto:info@pingcap.com">
         {{ partial "svgs/icon-mail.svg" (dict "width" 18 "height" 18 ) }}
         </a>
-        <a id="wechat_share_btn" data-site="wechat" data-url="{{ .URL | absURL }}" class="ssba_wechat_share" href="javascript:void(0)" onclick="toggleBlogQRCode()">
+        <a id="wechat_share_btn" data-site="wechat" data-url="{{ .RelPermalink | absURL }}" class="ssba_wechat_share" href="javascript:void(0)" onclick="toggleBlogQRCode()">
         {{ partial "svgs/icon-weixin.svg" (dict "width" 18 "height" 18 ) }}
         </a>
     </div>
