+++ layouts/partials/archives/post-share.html
@@ -1,22 +1,22 @@
 <div class="ssba-wrap">
     <div class="ssba">
-        <a data-site="" class="ssba_twitter_share" href="http://twitter.com/share?url={{ .URL | absURL }}" target="_blank">
+        <a data-site="" class="ssba_twitter_share" href="http://twitter.com/share?url={{ .RelPermalink | absURL }}" target="_blank">
         <!-- {{ partial "svgs/twitter-icon.svg" (dict "width" 18 "height" 18 ) }} -->
             <img src="/images/svgs/twitter-icon.svg">
         </a>
-        <a data-site="reddit" class="ssba_reddit_share" href="http://reddit.com/submit?url={{ .URL | absURL }}&title={{ .Params.title }}" target="_blank">
+        <a data-site="reddit" class="ssba_reddit_share" href="http://reddit.com/submit?url={{ .RelPermalink | absURL }}&title={{ .Params.title }}" target="_blank">
         <!-- {{ partial "svgs/reddit-icon.svg" (dict "width" 18 "height" 18 ) }} -->
             <img src="/images/svgs/reddit-icon.svg">
         </a>
-        <a data-site="" class="ssba_facebook_share" href="http://www.facebook.com/sharer.php?u={{ .URL | absURL }}" target="_blank">
+        <a data-site="" class="ssba_facebook_share" href="http://www.facebook.com/sharer.php?u={{ .RelPermalink | absURL }}" target="_blank">
         <!-- {{ partial "svgs/facebook-icon.svg" (dict "width" 18 "height" 18 ) }} -->
             <img src="/images/svgs/facebook-icon.svg">
         </a>
-        <a data-site="linkedin" class="ssba_linkedin_share" href="http://www.linkedin.com/shareArticle?mini=true&url={{ .URL | absURL }}" target="_blank">
+        <a data-site="linkedin" class="ssba_linkedin_share" href="http://www.linkedin.com/shareArticle?mini=true&url={{ .RelPermalink | absURL }}" target="_blank">
         <!-- {{ partial "svgs/linkedin-icon.svg" (dict "width" 18 "height" 18 ) }} -->
             <img src="/images/svgs/linkedin-icon.svg">
         </a>
-        <a data-site="hackernews" class="ssba_hackernews_share" href="https://news.ycombinator.com/submitlink?u={{ .URL | absURL }}&t={{ .Params.title }}" target="_blank">
+        <a data-site="hackernews" class="ssba_hackernews_share" href="https://news.ycombinator.com/submitlink?u={{ .RelPermalink | absURL }}&t={{ .Params.title }}" target="_blank">
         <!-- {{ partial "svgs/hacker-news-icon.svg" (dict "width" 18 "height" 18 ) }} -->
             <img src="/images/svgs/hacker-news-icon.svg">
         </a>
