+++ layouts/partials/head/meta-seo.html
@@ -29,7 +29,7 @@
 <meta property="og:type" content="website">
 <meta property="og:title" content="{{ .Title }} | {{ $.Scratch.Get "ogTitle" }}">
 <meta property="og:description" content="{{ $.Scratch.Get "summary" }}">
-<meta property="og:url" content="{{ .Page.URL | absURL }}">
+<meta property="og:url" content="{{ .Page.Permalink | absURL }}">
 <meta property="og:site_name" content="MySQL at Scale. No more manual sharding">
 <meta property="og:image" content="{{ $.Scratch.Get "image" }}">
 <meta property="og:image:secure_url" content="{{ $.Scratch.Get "image" }}">
