+++ layouts/partials/navs/date-tree-nav.html
@@ -21,7 +21,7 @@
                         <ul>
                             {{- range .Pages -}}
                             <li class="leaf-child">
-                                <a href="{{ .URL }}"><i class="icon-file"></i>{{- .Title -}}</a>
+                                <a href="{{ .RelPermalink }}"><i class="icon-file"></i>{{- .Title -}}</a>
                             </li>
                             {{- end -}}
                         </ul>
