+++ layouts/partials/recruit/positions-tab-panel.html
@@ -2,12 +2,12 @@
     {{- if .name -}}
         {{- range .posts -}}
             {{- if in .Params.tags $.global.name -}}
-                {{- partial "recruit/card.html" (dict "avatar" .Params.avatar "title" .Params.title "location" .Params.location "link" .URL "season" .Params.season) -}}
+                {{- partial "recruit/card.html" (dict "avatar" .Params.avatar "title" .Params.title "location" .Params.location "link" .RelPermalink "season" .Params.season) -}}
             {{- end -}}
         {{- end -}}
     {{- else -}}
         {{- range .posts -}}
-            {{- partial "recruit/card.html" (dict "avatar" .Params.avatar "title" .Params.title "location" .Params.location "link" .URL "season" .Params.season) -}}
+            {{- partial "recruit/card.html" (dict "avatar" .Params.avatar "title" .Params.title "location" .Params.location "link" .RelPermalink "season" .Params.season) -}}
         {{- end -}}
     {{- end -}}
 </div>
