+++ package.json
@@ -5,85 +5,54 @@
   "main": "app.js",
   "scripts": {
     "hugo": "gulp hugo",
-    "gen-content": "bash ./scripts/gen-sub-gitinfo.sh;bash ./scripts/gen-content.sh",
-    "clean-build": "rm -rf dist;",
-    "build": "git submodule sync;git submodule update --init --recursive --remote;npm run gen-content;export NODE_ENV=production;gulp css;./scripts/build-check-err.sh",
-    "build-preview": "gulp build-preview",
-    "start": "export NODE_ENV=development;gulp server",
+    "gen-content": "bash ./scripts/gen-sub-gitinfo.sh; bash ./scripts/gen-content.sh",
+    "clean-build": "rimraf dist",
+    "build": "git submodule sync; git submodule update --init --recursive --remote; npm run gen-content; export NODE_ENV=production; ./scripts/build-check-err.sh",
+    "build-preview": "gulp buildPreview",
+    "start": "export NODE_ENV=development; gulp server",
     "lint": "eslint src"
   },
   "author": "",
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
+    "@babel/core": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
     "@octokit/rest": "^15.11.0",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^7.0.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-register": "^6.11.6",
+    "babel-loader": "^8.0.6",
     "body-parser": "^1.18.3",
-    "browser-sync": "^2.24.4",
+    "browser-sync": "^2.26.7",
     "css-loader": "^0.28.7",
     "cssnano": "^3.10.0",
-    "eslint": "^4.7.2",
-    "eslint-config-prettier": "^2.6.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-prettier": "^2.3.1",
+    "eslint": "^6.1.0",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-prettier": "^3.1.0",
     "exports-loader": "^0.6.3",
     "extract-text-webpack-plugin": "^3.0.1",
+    "fancy-log": "^1.3.3",
     "file-loader": "^0.11.1",
-    "gulp": "^3.9.1",
+    "gulp": "^4.0.2",
     "gulp-autoprefixer": "^5.0.0",
-    "gulp-babel": "^6.1.2",
     "gulp-clean-css": "^3.9.4",
     "gulp-less": "^3.3.2",
-    "gulp-util": "^3.0.7",
     "html-webpack-plugin": "^2.30.1",
-    "hugo-bin": "^0.17.0",
+    "hugo-bin": "0.44.3",
     "imports-loader": "^0.7.1",
     "json-server": "^0.14.0",
     "lodash": "^4.17.11",
     "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "plugin-error": "^1.0.1",
     "postcss-cssnext": "^2.7.0",
     "postcss-import": "^10.0.0",
     "preload-webpack-plugin": "^2.3.0",
-    "prettier": "^1.7.0",
+    "prettier": "^1.18.2",
+    "rimraf": "^2.6.3",
     "stylelint": "^9.2.1",
     "stylelint-config-standard": "^18.2.0",
     "url-loader": "^0.5.7",
     "webpack": "^3.8.1",
     "whatwg-fetch": "^1.0.0"
-  },
-  "babel": {
-    "presets": [
-      "env"
-    ]
-  },
-  "eslintConfig": {
-    "extends": "prettier",
-    "plugins": [
-      "prettier",
-      "import"
-    ],
-    "parserOptions": {
-      "ecmaVersion": 2016,
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "jsx": true
-      }
-    },
-    "rules": {
-      "prettier/prettier": [
-        "error",
-        {
-          "singleQuote": true,
-          "semi": false,
-          "trailingComma": "es5",
-          "jsxBracketSameLine": true
-        }
-      ],
-      "no-script-url": 0
-    }
   },
   "stylelint": {
     "extends": "stylelint-config-standard",
